requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
langchain
langchain-google-genai
chromadb
flask
python-dotenv
pypdf
python-docx
markdown 